Fix compile warnings.
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/libssh/trunk@276 7dcaeef0-15fb-0310-b436-a5af3365683c

+++ libssh/sftp.c
@@ -648,7 +648,9 @@ static SFTP_ATTRIBUTES *sftp_parse_attr_3(SFTP_SESSION *sftp, BUFFER *buf,
             if(buffer_get_u64(buf,&attr->size)!=sizeof(u64))
                 break;
             attr->size=ntohll(attr->size);
-            ssh_log(sftp->session, SSH_LOG_RARE, "Size: %lu\n", attr->size);
+            ssh_log(sftp->session, SSH_LOG_RARE,
+                "Size: %llu\n",
+                (long long unsigned int) attr->size);
         }
         if(flags & SSH_FILEXFER_ATTR_UIDGID){
             if(buffer_get_u32(buf,&attr->uid)!=sizeof(u32))
